added file browser and stubbed key generation
