+++ README.md
@@ -1,0 +1,11 @@
+# Assignment 2
+
+## Task 1
+
+## Task 2
+
+## Task 3
+
+## Task 4
+To run: `python task4`.
+Task 4 requires a gui. The library is tkinter which should be included in your python install, no extra installation required.
