+++ task4/key_gen.py
@@ -1,0 +1,6 @@
+
+
+
+def generate_keys():
+    # TODO: Actually generate the keys here
+    return ('public_key', 'private_key')
